refactoring name of extractors
